@@ -3,29 +3,29 @@
   "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.21",
-    "@fortawesome/free-solid-svg-icons": "5.10.1",
-    "@fortawesome/react-fontawesome": "0.1.4",
+    "@fortawesome/fontawesome-svg-core": "1.2.25",
+    "@fortawesome/free-solid-svg-icons": "5.11.2",
+    "@fortawesome/react-fontawesome": "0.1.5",
     "bugsnag-js": "4.7.3",
     "bugsnag-react": "^1.1.1",
     "classnames": "^2.2.6",
     "core-js": "^3.2.0",
     "create-react-class": "^15.6.3",
     "firebase": "6.3.5",
-    "firebaseui": "4.1.0",
+    "firebaseui": "4.2.0",
     "jaro-winkler": "^0.2.8",
     "lodash-es": "4.17.15",
     "notie": "^4.3.1",
     "prop-types": "15.7.2",
-    "react": "16.9.0",
-    "react-dom": "16.9.0",
-    "react-ga": "2.6.0",
+    "react": "16.10.2",
+    "react-dom": "16.10.2",
+    "react-ga": "2.7.0",
     "react-highlight-words": "0.16.0",
     "react-hotkeys": "2.0.0",
-    "react-modal": "3.9.1",
+    "react-modal": "3.10.1",
     "react-paypal-button-v2": "^2.3.2",
     "react-pdf-js": "1",
-    "react-router-dom": "5.0.1",
+    "react-router-dom": "5.1.2",
     "react-select": "^3.0.4",
     "react-virtualized": "9.21.1",
     "reactstrap": "8.0.1",
@@ -33,7 +33,7 @@
     "whatwg-fetch": "3.0.0"
   },
   "devDependencies": {
-    "react-scripts": "3.1.0",
+    "react-scripts": "3.2.0",
     "react-snap": "^1.23.0",
     "redux-devtools": "3.5.0",
     "redux-devtools-dock-monitor": "1.1.3",
